requirements.txt
streamlit>=1.28.0
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.10
langchain-openai>=0.0.5
langgraph>=0.0.20
python-dotenv>=1.0.0 